REFACTOR: flag, transport / service submodules
