+++ core/internal/app/flag/service/flag_service.go
@@ -1,4 +1,4 @@
-package flag
+package service
 
 import (
 	"context"
