+++ core/internal/app/flag/service/flag_util.go
@@ -1,4 +1,4 @@
-package flag
+package service
 
 import (
 	"core/internal/app/environment"
